Allow for searching nested dice
fixes damage rolls in PF2e not being triggered. #28

Adds an optional configuration to search for nested dice. With it is turned on, for each `roll` we’ll not only look if the term is a `Die` but also see if the non-die term answers to `term.dice`, in which case we’ll recursively look for the dice in there.

This solves issues in the cases where we have special Roll types that don’t conform usual pattern of a straight formula.

An example is a damage roll in Pathfinder 2e (or spell damage, or crit…) where without this option enabled, `unfulfilled-rolls` will not be able to act on them.

Also adding `.history` to gitignore to avoid issues with VS Code history plugin.